requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pyaes==1.6.1
pyasn1==0.6.1
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.1
rsa==4.9
schedule==1.2.2
Telethon==1.39.0
